changed how we are exporting pagination_renderer.rb
